Split tagged/discriminated union tests out of test_union.py
`test_field_serializer_in_nested_tagged_union_called_only_twice` and
`test_union_tags_in_errors` move to test_discriminated_union.py, which
covers tagged/discriminated unions; test_union.py now only holds plain
(smart/left-to-right) union tests. No duplicate tests were found
between the two files.
